Move validate database/messaging commands to status
Rename the validate database and validate messaging commands to
status database and status messaging respectively. The commands
check the cluster status of MariaDB Galera and RabbitMQ and fit
better under the status namespace.

AI-assisted: Claude Code
