+++ setup.cfg
@@ -147,9 +147,9 @@ osism.commands:
     lock = osism.commands.lock:Lock
     unlock = osism.commands.lock:Unlock
     lock status = osism.commands.lock:LockStatus
+    status database = osism.commands.status:Database
+    status messaging = osism.commands.status:Messaging
     validate = osism.commands.validate:Run
-    validate database = osism.commands.validate:Database
-    validate messaging = osism.commands.validate:Messaging
     vault password set = osism.commands.vault:SetPassword
     vault password unset = osism.commands.vault:UnsetPassword
     vault view = osism.commands.vault:View
